Changed all floats to double, made plane kinda working, lightning/shadows are kinda wonky still

+++ RayTracer/src/RayTracer18/Light/PointLight.java
@@ -5,7 +5,7 @@ import javafx.scene.paint.Color;
 
 public class PointLight extends Light{
 
-    public PointLight(Vector3 position, float intensity, Color color) {
+    public PointLight(Vector3 position, double intensity, Color color) {
         super(position, intensity, color);
     }
 }
